apps/open-source/token-server/requirements.txt
fastapi
uvicorn
python-dotenv
livekit-api
supabase
pydantic[email]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib